requirements.txt
moviepy
gTTS
mutagen
git+https://github.com/tweepy/tweepy.git
pandas
wget
pytesseract
pillow
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
httplib2
oauth2client 
